% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rawdata}
\alias{rawdata}
\title{dataset containing coolness scores for students in HEV90009 PhD course}
\format{
\subsection{\code{rawdata}}{

A data frame with 5 rows and 2 columns:
\describe{
\item{student}{student names}
\item{coolness}{coolness score; 0-100}
}
}
}
\usage{
rawdata
}
\description{
This dataset is important
}
\keyword{datasets}
